+++ routes/auth_routes.py
@@ -9,6 +9,7 @@ from flask_mail import Message # type: ignore
 import random
 from datetime import datetime, timedelta  
 import pytz
+import routes.dashboard_routes as dashboard_routes
 
 auth_bp = Blueprint('auth_bp', __name__, url_prefix='/auth')
 
@@ -56,7 +57,7 @@ def login():
 
                 if user.has_role('admin'):
                     # Redirecion al dashboard
-                    return redirect(url_for('personal_bp.dashboard'))
+                    return redirect(url_for('dashboard_bp.index'))
                 return redirect(url_for('cliente_bp.index'))
 
             else:
