+++ templates/dashboard/dashboard.html
@@ -1,85 +1,95 @@
 {% extends "layout.html" %}
 
 {% block content %}
-<script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-
-<div class="flex flex-col w-full px-4 md:px-10">
-    <div class="w-full p-4 text-center bg-white border border-gray-200 rounded-lg shadow-sm">
-        <!-- Título INVENTARIO centrado -->
-        <h1 class="text-3xl font-bold text-center mb-6">Dashboard de Galletas y Ventas</h1>
-
-        <!-- Sección de Gráficas -->
-        <div class="grid grid-cols-1 md:grid-cols-2 gap-6">
-
-            <!-- Gráfica de Ventas -->
-            <div class="bg-white p-6 rounded-lg shadow-xl">
-                <h2 class="text-xl font-semibold mb-4">Ventas Mensuales</h2>
-                <canvas id="ventasChart" class="w-full h-72"></canvas>
-            </div>
-
-            <!-- Gráfica de Galletas -->
-            <div class="bg-white p-6 rounded-lg shadow-xl">
-                <h2 class="text-xl font-semibold mb-4">Galletas Vendidas</h2>
-                <canvas id="galletasChart" class="w-full h-72"></canvas>
-            </div>
-
+<div class="flex flex-col w-full px-10">
+  <!-- Encabezado del Dashboard -->
+  <div class="w-full p-4 border border-gray-300 rounded-lg shadow-sm mb-6">
+    <div class="flex flex-col pb-5">
+      <div>
+        <h1 class="text-4xl font-medium text-gray-800 mb-4">Dashboard</h1>
+      </div>
+      <!-- Sección de Métricas Principales -->
+      <div class="flex flex-col lg:flex-row gap-4 mb-6">
+        <!-- Ventas del Día (más grande visualmente, no ocupa dos columnas) -->
+        <div class="flex-1 p-6 bg-green-50 border-2 border-green-300 rounded-2xl shadow-md text-center scale-105">
+          <h2 class="text-xl font-semibold text-green-800 mb-2">Ventas del Día</h2>
+          <p class="text-4xl font-extrabold text-green-700">${{ ventas_dia }}</p>
         </div>
+
+        <!-- Compras del Día -->
+        <div class="flex-1 p-6 bg-red-50 border border-red-200 rounded-2xl shadow-sm text-center">
+          <h2 class="text-lg font-semibold text-red-800 mb-1">Compras del Día</h2>
+          <p class="text-3xl font-bold text-red-700">${{ compras_dia }}</p>
+        </div>
+
+        <!-- Fondo Inicial -->
+        <div class="flex-1 p-6 bg-blue-50 border border-blue-200 rounded-2xl shadow-sm text-center">
+          <h2 class="text-lg font-semibold text-blue-800 mb-1">Fondo Inicial</h2>
+          <p class="text-3xl font-bold text-blue-700">${{ fondo_inicial }}</p>
+        </div>
+
+        <!-- Total en Caja -->
+        <div class="flex-1 p-6 bg-yellow-50 border border-yellow-200 rounded-2xl shadow-sm text-center">
+          <h2 class="text-lg font-semibold text-yellow-800 mb-1">Total en Caja</h2>
+          <p class="text-3xl font-bold text-yellow-700">${{ total_caja }}</p>
+        </div>
+      </div>
+
+      <!-- Top 5 Galletas Más Vendidas -->
+      <div class="mb-6">
+        <h2 class="text-2xl font-semibold text-gray-800 mb-4">Top 5 Galletas Más Vendidas 🏆</h2>
+        <div class="flex gap-6 overflow-x-auto">
+          {% for galleta in top_5_galletas %}
+          <div class="p-4 bg-white border border-gray-200 rounded-lg shadow-md flex flex-col items-center w-1/5">
+            <img
+            src="{{ url_for('galletas_bp.mostrar_imagen', galleta_id=galleta.id) }}"
+            class="h-32 w-32 object-cover rounded-lg mb-4"
+            alt="{{ galleta.nombre }}"
+          />          
+            <h3 class="text-xl font-semibold text-gray-700 text-center">{{ galleta.nombre }}</h3>
+          </div>
+          {% endfor %}
+        </div>
+      </div>
+
+      <!-- Inventario Bajo / Crítico -->
+      <div class="mb-6">
+        <h2 class="text-2xl font-semibold text-red-800 mb-4">Inventario Bajo / Crítico</h2>
+        <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6">
+          {% for insumo in inventario_critico %}
+          <div class="p-4 bg-yellow-50 border border-yellow-300 rounded-lg shadow-md">
+            <h3 class="text-xl font-semibold text-yellow-800">{{ insumo.nombre }}</h3>
+            <p class="text-lg font-medium text-red-600">⚠️ {{ insumo.cantidad_restante }} restantes</p>
+          </div>
+          {% endfor %}
+          {% if not inventario_critico %}
+          <p class="text-center text-lg text-gray-600">No hay insumos críticos.</p>
+          {% endif %}
+        </div>
+      </div>
+
+      <!-- Pedidos para Hoy -->
+      <div class="mb-6">
+        <h2 class="text-2xl font-semibold text-blue-800 mb-4">Pedidos para Hoy</h2>
+        {% if pedidos_hoy %}
+        <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6">
+          {% for pedido in pedidos_hoy %}
+          <div class="p-4 bg-white border border-gray-200 rounded-lg shadow-md">
+            <h3 class="text-xl font-semibold text-gray-700">{{ pedido.cliente_nombre }}</h3>
+            <p class="text-lg font-medium text-green-600">Pedido: {{ pedido.producto }}</p>
+            <p class="text-lg font-medium text-gray-600">Fecha de entrega: {{ pedido.fecha_entrega }}</p>
+          </div>
+          {% endfor %}
+        </div>
+        {% else %}
+        <p class="text-center text-lg text-gray-600">No hay pedidos para hoy.</p>
+        {% endif %}
+      </div>
     </div>
+  </div>
 </div>
 
-<!-- Script de Chart.js -->
-<script>
-    // Datos Ficticios para Ventas Mensuales
-    var ventasData = {
-        labels: ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio'],
-        datasets: [{
-            label: 'Ventas en USD',
-            data: [1000, 1500, 1200, 1800, 2500, 2200],
-            backgroundColor: 'rgba(75, 192, 192, 0.2)',
-            borderColor: 'rgba(75, 192, 192, 1)',
-            borderWidth: 1
-        }]
-    };
-
-    // Gráfica de Ventas
-    var ctxVentas = document.getElementById('ventasChart').getContext('2d');
-    var ventasChart = new Chart(ctxVentas, {
-        type: 'line',  // Tipo de gráfico
-        data: ventasData,
-        options: {
-            scales: {
-                y: {
-                    beginAtZero: true
-                }
-            }
-        }
-    });
-
-    // Datos Ficticios para Galletas Vendidas
-    var galletasData = {
-        labels: ['Galleta A', 'Galleta B', 'Galleta C', 'Galleta D', 'Galleta E'],
-        datasets: [{
-            label: 'Unidades Vendidas',
-            data: [200, 300, 150, 400, 500],
-            backgroundColor: 'rgba(255, 159, 64, 0.2)',
-            borderColor: 'rgba(255, 159, 64, 1)',
-            borderWidth: 1
-        }]
-    };
-
-    // Gráfica de Galletas Vendidas
-    var ctxGalletas = document.getElementById('galletasChart').getContext('2d');
-    var galletasChart = new Chart(ctxGalletas, {
-        type: 'bar',  // Tipo de gráfico
-        data: galletasData,
-        options: {
-            scales: {
-                y: {
-                    beginAtZero: true
-                }
-            }
-        }
-    });
-</script>
+<!-- Incluye Flowbite JS -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/flowbite/2.3.0/flowbite.min.js"></script>
 
 {% endblock %}
