--- templates/dashboard_joel/dashboard.html
@@ -1,97 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-<div class="flex flex-col w-full px-10">
-  <!-- Encabezado del Dashboard -->
-  <div class="w-full p-4 border border-gray-300 rounded-lg shadow-sm mb-6">
-    <div class="flex flex-col pb-5">
-      <div>
-        <h1 class="text-4xl font-medium text-gray-800 mb-4">Dashboard</h1>
-      </div>
-      <!-- Sección de Métricas Principales -->
-      <div class="flex flex-col lg:flex-row gap-4 mb-6">
-        <!-- Ventas del Día (más grande visualmente, no ocupa dos columnas) -->
-        <div class="flex-1 p-6 bg-green-50 border-2 border-green-300 rounded-2xl shadow-md text-center scale-105">
-          <h2 class="text-xl font-semibold text-green-800 mb-2">Ventas del Día</h2>
-          <p class="text-4xl font-extrabold text-green-700">${{ ventas_dia }}</p>
-        </div>
-
-        <!-- Compras del Día -->
-        <div class="flex-1 p-6 bg-red-50 border border-red-200 rounded-2xl shadow-sm text-center">
-          <h2 class="text-lg font-semibold text-red-800 mb-1">Compras del Día</h2>
-          <p class="text-3xl font-bold text-red-700">${{ compras_dia }}</p>
-        </div>
-
-        <!-- Fondo Inicial -->
-        <div class="flex-1 p-6 bg-blue-50 border border-blue-200 rounded-2xl shadow-sm text-center">
-          <h2 class="text-lg font-semibold text-blue-800 mb-1">Fondo Inicial</h2>
-          <p class="text-3xl font-bold text-blue-700">${{ fondo_inicial }}</p>
-        </div>
-
-        <!-- Total en Caja -->
-        <div class="flex-1 p-6 bg-yellow-50 border border-yellow-200 rounded-2xl shadow-sm text-center">
-          <h2 class="text-lg font-semibold text-yellow-800 mb-1">Total en Caja</h2>
-          <p class="text-3xl font-bold text-yellow-700">${{ total_caja }}</p>
-        </div>
-      </div>
-
-      <!-- Top 5 Galletas Más Vendidas -->
-      <div class="mb-6">
-        <h2 class="text-2xl font-semibold text-gray-800 mb-4">Top 5 Galletas Más Vendidas 🏆</h2>
-        <div class="flex gap-6 overflow-x-auto">
-          {% for galleta in top_5_galletas %}
-          <div class="p-4 bg-white border border-gray-200 rounded-lg shadow-md flex flex-col items-center w-1/5">
-            <!-- Redimensionar la imagen -->
-            <img
-              src="{{ url_for('galletas_bp.mostrar_imagen', galleta_id=galleta.id) }}"
-              class="h-32 w-32 object-cover rounded-lg mb-4"
-              alt="{{ galleta.nombre }}"
-            />
-            <h3 class="text-xl font-semibold text-gray-700 text-center">{{ galleta.nombre }}</h3>
-            <p class="text-lg font-medium text-green-600 text-center">{{ galleta.total_vendido }} ventas</p>
-          </div>
-          {% endfor %}
-        </div>
-      </div>
-
-      <!-- Inventario Bajo / Crítico -->
-      <div class="mb-6">
-        <h2 class="text-2xl font-semibold text-red-800 mb-4">Inventario Bajo / Crítico</h2>
-        <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6">
-          {% for insumo in inventario_critico %}
-          <div class="p-4 bg-yellow-50 border border-yellow-300 rounded-lg shadow-md">
-            <h3 class="text-xl font-semibold text-yellow-800">{{ insumo.nombre }}</h3>
-            <p class="text-lg font-medium text-red-600">⚠️ {{ insumo.cantidad_restante }} restantes</p>
-          </div>
-          {% endfor %}
-          {% if not inventario_critico %}
-          <p class="text-center text-lg text-gray-600">No hay insumos críticos.</p>
-          {% endif %}
-        </div>
-      </div>
-
-      <!-- Pedidos para Hoy -->
-      <div class="mb-6">
-        <h2 class="text-2xl font-semibold text-blue-800 mb-4">Pedidos para Hoy</h2>
-        {% if pedidos_hoy %}
-        <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6">
-          {% for pedido in pedidos_hoy %}
-          <div class="p-4 bg-white border border-gray-200 rounded-lg shadow-md">
-            <h3 class="text-xl font-semibold text-gray-700">{{ pedido.cliente_nombre }}</h3>
-            <p class="text-lg font-medium text-green-600">Pedido: {{ pedido.producto }}</p>
-            <p class="text-lg font-medium text-gray-600">Fecha de entrega: {{ pedido.fecha_entrega }}</p>
-          </div>
-          {% endfor %}
-        </div>
-        {% else %}
-        <p class="text-center text-lg text-gray-600">No hay pedidos para hoy.</p>
-        {% endif %}
-      </div>
-    </div>
-  </div>
-</div>
-
-<!-- Incluye Flowbite JS -->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/flowbite/2.3.0/flowbite.min.js"></script>
-
-{% endblock %}
